refactor: 프로덕션 코드를 리팩토링한다
