+++ src/main/java/shopping/domain/Product.java
@@ -1,14 +1,12 @@
 package shopping.domain;
 
-import shopping.exception.ProductException;
-import shopping.utils.StringUtils;
-
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import java.util.Objects;
+import shopping.exception.ProductException;
+import shopping.utils.StringUtils;
 
 @Entity
 public class Product {
