+++ src/main/java/shopping/infrastructure/CurrencyManager.java
@@ -1,11 +1,13 @@
-package shopping.currency;
+package shopping.infrastructure;
 
 import java.text.MessageFormat;
 import java.util.Objects;
 import java.util.Optional;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
+import shopping.dto.CurrencyLayerResponse;
 import shopping.exception.CurrencyLayerException;
+import shopping.infrastructure.ApiService;
 
 @Service
 public class CurrencyManager {
